requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
tqdm==4.65.0
PyYAML==6.0.1
